dynamic exercise pagewith custom 404 page
