+++ src/containers/exerciseView/exerciseView.stories.tsx
@@ -1,19 +1,14 @@
 import { StoryObj, Meta } from '@storybook/react';
 import ExerciseView from '.';
 
-import { crossfitExercises } from '../../server/crossfitExercises/exerciseExamples';
-
 export default {
   title: 'Views/ExerciseView',
   component: ExerciseView,
-  args: {
-    crossfitExercises,
-  },
+  args: {},
 } as Meta;
 
 type Story = StoryObj<typeof ExerciseView>;
 
 export const Default: Story = {
-  args: {
-  },
+  args: {},
 };
